Add color picker tab preference
