+++ lib/core/preferences/layout_preferences.dart
@@ -9,4 +9,8 @@ class LayoutPreferences {
   /// timeline rather than being accessible through a settings screen.
   bool get mediaTiled => harpyPrefs.getBool('mediaTiled', true);
   set mediaTiled(bool value) => harpyPrefs.setBool('mediaTiled', value);
+
+  /// The id of the tab that is used by the harpy color picker.
+  int get colorPickerTab => harpyPrefs.getInt('colorPickerTab', 0);
+  set colorPickerTab(int value) => harpyPrefs.setInt('colorPickerTab', value);
 }
